refactor: add custom hook about fileTree & fileTab
